# cmd/swagger/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "飞宏IM接口",
        "title": "飞宏IM接口",
        "contact": {
            "name": "飞宏IM技术支持",
            "url": "https://github.com/liony823/chat",
            "email": "ilovecoding@foxmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "host": "localhost:10002",
    "basePath": "/",
    "paths": {
...
        "/object/auth_sign": {
            "post": {
                "description": "安全签名",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "object"
                ],
                "summary": "安全签名",
                "operationId": "authSign",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/third.AuthSignReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/apiresp.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/third.AuthSignResp"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/object/complete_multipart_upload": {
            "post": {
                "description": "完成上传",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "object"
                ],
                "summary": "完成上传",
                "operationId": "completeMultipartUpload",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/third.CompleteMultipartUploadReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/apiresp.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/third.CompleteMultipartUploadResp"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/object/initiate_multipart_upload": {
            "post": {
                "description": "初始化上传",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "object"
                ],
                "summary": "初始化上传",
                "operationId": "initiateMultipartUpload",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/third.InitiateMultipartUploadReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/apiresp.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/third.InitiateMultipartUploadResp"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/object/part_limit": {
            "post": {
                "description": "获取分片范围",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "object"
                ],
                "summary": "获取分片范围",
                "operationId": "partLimit",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/third.PartLimitReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/apiresp.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/third.PartLimitResp"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/object/part_size": {
            "post": {
                "description": "获取分片大小",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "object"
                ],
                "summary": "获取分片大小",
                "operationId": "partSize",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/third.PartSizeReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/apiresp.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/third.PartSizeResp"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "apiresp.ApiResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "errCode": {
                    "type": "integer"
                },
                "errDlt": {
                    "type": "string"
                },
                "errMsg": {
                    "type": "string"
                }
            }
        },
...
        "third.AuthSignParts": {
            "type": "object",
            "properties": {
                "header": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/third.KeyValues"
                    }
                },
                "parts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/third.SignPart"
                    }
                },
                "query": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/third.KeyValues"
                    }
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "third.AuthSignReq": {
            "type": "object",
            "properties": {
                "partNumbers": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "uploadID": {
                    "type": "string"
                }
            }
        },
        "third.AuthSignResp": {
            "type": "object",
            "properties": {
                "header": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/third.KeyValues"
                    }
                },
                "parts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/third.SignPart"
                    }
                },
                "query": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/third.KeyValues"
                    }
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "third.CompleteMultipartUploadReq": {
            "type": "object",
            "properties": {
                "cause": {
                    "type": "string"
                },
                "contentType": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "parts": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "uploadID": {
                    "type": "string"
                },
                "urlPrefix": {
                    "type": "string"
                }
            }
        },
        "third.CompleteMultipartUploadResp": {
            "type": "object",
            "properties": {
                "url": {
                    "type": "string"
                }
            }
        },
        "third.InitiateMultipartUploadReq": {
            "type": "object",
            "properties": {
                "cause": {
                    "type": "string"
                },
                "contentType": {
                    "type": "string"
                },
                "hash": {
                    "type": "string"
                },
                "maxParts": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "partSize": {
                    "type": "integer"
                },
                "size": {
                    "type": "integer"
                },
                "urlPrefix": {
                    "type": "string"
                }
            }
        },
        "third.InitiateMultipartUploadResp": {
            "type": "object",
            "properties": {
                "upload": {
                    "$ref": "#/definitions/third.UploadInfo"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "third.KeyValues": {
            "type": "object",
            "properties": {
                "key": {
                    "type": "string"
                },
                "values": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "third.PartLimitReq": {
            "type": "object"
        },
        "third.PartLimitResp": {
            "type": "object",
            "properties": {
                "maxNumSize": {
                    "type": "integer"
                },
                "maxPartSize": {
                    "type": "integer"
                },
                "minPartSize": {
                    "type": "integer"
                }
            }
        },
        "third.PartSizeReq": {
            "type": "object",
            "properties": {
                "size": {
                    "type": "integer"
                }
            }
        },
        "third.PartSizeResp": {
            "type": "object",
            "properties": {
                "size": {
                    "type": "integer"
                }
            }
        },
        "third.SignPart": {
            "type": "object",
            "properties": {
                "header": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/third.KeyValues"
                    }
                },
                "partNumber": {
                    "type": "integer"
                },
                "query": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/third.KeyValues"
                    }
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "third.UploadInfo": {
            "type": "object",
            "properties": {
                "expireTime": {
                    "type": "integer"
                },
                "partSize": {
                    "type": "integer"
                },
                "sign": {
                    "$ref": "#/definitions/third.AuthSignParts"
                },
                "uploadID": {
                    "type": "string"
                }
            }
        },
...
    }
}